requirements.txt
appdirs==1.4.0
Django==1.10
django-braces==1.10.0
gnureadline==6.3.3
ipython==3.0.0
packaging==16.8
pyparsing==2.1.10
six==1.10.0
